Movies api crud implementation

+++ backend/src/genres/genres.controller.ts
@@ -9,26 +9,38 @@ export class GenresController {
 
   @Post()
   create(@Body() createGenreDto: CreateGenreDto) {
-    return this.genresService.create(createGenreDto);
+    if('tmdb_id' in createGenreDto && 'name' in createGenreDto){
+      return {data:this.genresService.create(createGenreDto), success:true};
+    }
+    else  return {success:false, error:"All fields are required"}
   }
 
   @Get()
   findAll() {
-    return this.genresService.findAll();
+    return {data:this.genresService.findAll(), success:true};
   }
 
   @Get(':id')
   findOne(@Param('id') id: string) {
-    return this.genresService.findOne(+id);
+    if (!id.match(/^[0-9a-fA-F]{24}$/)) return { error: 'Invalid id',success: false};
+    const genre = this.genresService.findOne(+id);
+    if (genre) return {data:genre, success:true};
+    return { error: 'Movie genre not found', success:false};
   }
 
   @Patch(':id')
   update(@Param('id') id: string, @Body() updateGenreDto: UpdateGenreDto) {
-    return this.genresService.update(+id, updateGenreDto);
+    if (!id.match(/^[0-9a-fA-F]{24}$/)) return { error: 'Invalid id', success: false};
+    const genre = this.genresService.update(+id, updateGenreDto);
+    if (genre) return {data:genre, success:true};
+    return { error: 'Concert genre not found', success:false };
   }
 
   @Delete(':id')
   remove(@Param('id') id: string) {
-    return this.genresService.remove(+id);
+    if (!id.match(/^[0-9a-fA-F]{24}$/)) return { error: 'Invalid id', success: false};
+    const genre = this.genresService.remove(+id);
+    if (genre) return {data:genre, success:true};
+    return { error: 'Movie genre not found', success:false};
   }
 }
